backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
psycopg2-binary==2.9.9
gunicorn==22.0.0